requirements.txt
langchain
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
python-dotenv
sentence-transformers